remove independent datastack backend, use plotting backend instead, remove broken plotting functions
Remove removed subpackage from setup.py
fix more instances that imported list this: "from sherpa.astro.plot import backend"
