+++ setup.py
@@ -89,7 +89,6 @@ meta = dict(name='sherpa',
                       'sherpa.utils',
                       'sherpa.astro',
                       'sherpa.astro.datastack',
-                      'sherpa.astro.datastack.plot_backend',
                       'sherpa.astro.io',
                       'sherpa.astro.models',
                       'sherpa.astro.optical',
