+++ sherpa/astro/data.py
@@ -283,7 +283,8 @@ def _extract_fields(obj, summary):
 def html_pha(pha):
     """HTML representation: PHA"""
 
-    from sherpa.astro.plot import DataPHAPlot, backend
+    from sherpa.astro.plot import DataPHAPlot
+    from sherpa import plot
 
     ls = []
 
@@ -291,7 +292,7 @@ def html_pha(pha):
     plotter.prepare(pha)
 
     try:
-        out = backend.as_html_plot(plotter, 'PHA Plot')
+        out = plot.backend.as_html_plot(plotter, 'PHA Plot')
     except AttributeError:
         out = None
 
@@ -462,7 +463,8 @@ def html_arf(arf):
     # It also assumes the units are keV/cm^2 which is not
     # guaranteed.
 
-    from sherpa.astro.plot import ARFPlot, backend
+    from sherpa.astro.plot import ARFPlot
+    from sherpa import plot
 
     ls = []
 
@@ -470,7 +472,7 @@ def html_arf(arf):
     plotter.prepare(arf)
 
     try:
-        out = backend.as_html_plot(plotter, 'ARF Plot')
+        out = plot.backend.as_html_plot(plotter, 'ARF Plot')
     except AttributeError:
         out = None
 
@@ -688,7 +690,7 @@ def simulate_rmf_plot(rmf):
     """
 
     from sherpa.models.basic import Delta1D
-    from sherpa.plot import backend
+    from sherpa import plot
 
     try:
         from matplotlib import pyplot as plt
@@ -740,7 +742,7 @@ def simulate_rmf_plot(rmf):
         return fig
 
     try:
-        return backend.as_svg(plotfunc)
+        return plot.backend.as_svg(plotfunc)
     except AttributeError:
         return None
 
@@ -756,7 +758,7 @@ def img_plot(img):
 
     """
 
-    from sherpa.plot import backend
+    from sherpa import plot
 
     try:
         from matplotlib import pyplot as plt
@@ -818,7 +820,7 @@ def img_plot(img):
         return fig
 
     try:
-        return backend.as_svg(plotfunc)
+        return plot.backend.as_svg(plotfunc)
     except AttributeError:
         return None
 
@@ -1078,8 +1080,8 @@ class DataARF(DataOgipResponse):
         return 'Energy (keV)'
 
     def get_ylabel(self):
-        from sherpa.plot import backend
-        return 'cm' + backend.get_latex_for_string('^2')
+        from sherpa import plot
+        return 'cm' + plot.backend.get_latex_for_string('^2')
 
 
 class DataRMF(DataOgipResponse):
@@ -3740,9 +3742,9 @@ must be an integer.""")
 
         if self.plot_fac:
             from sherpa.plot import backend
-            latex = backend.get_latex_for_string(f'^{self.plot_fac}')
-            ylabel += f' X {self.units.capitalize()}{latex}'
-
+            latex = backend.get_latex_for_string(
+                '^{}'.format(self.plot_fac))
+            ylabel += ' X {}{}'.format(self.units.capitalize(), latex)
         return ylabel
 
     @staticmethod
